# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RAFMARKET - Pi Network Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PiAuthInput {
  accessToken: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
  seller: 'seller',
} as const;

export interface User {
  id: number;
  piUsername: string;
  piUid?: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface CategoryPatch {
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  active: 'active',
  inactive: 'inactive',
  sold_out: 'sold_out',
} as const;

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  currency: string;
  stock: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  sellerId: number;
  /** @nullable */
  sellerUsername?: string | null;
  imageUrls: string[];
  tags?: string[];
  status: ProductStatus;
  isFeatured?: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  currency: string;
  stock: number;
  categoryId: number;
  imageUrls?: string[];
  tags?: string[];
  isFeatured?: boolean;
}

export type ProductPatchStatus = typeof ProductPatchStatus[keyof typeof ProductPatchStatus];


export const ProductPatchStatus = {
  active: 'active',
  inactive: 'inactive',
  sold_out: 'sold_out',
} as const;

export interface ProductPatch {
  name?: string;
  description?: string;
  price?: number;
  currency?: string;
  stock?: number;
  categoryId?: number;
  imageUrls?: string[];
  tags?: string[];
  status?: ProductPatchStatus;
  isFeatured?: boolean;
}

export interface ProductPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  currency: string;
  quantity: number;
  /** @nullable */
  imageUrl: string | null;
}

export interface Cart {
  items: CartItem[];
  total: number;
  currency: string;
  itemCount?: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemPatch {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  currency: string;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  buyerUsername?: string | null;
  items: OrderItem[];
  total: number;
  currency: string;
  status: OrderStatus;
  /** @nullable */
  piPaymentId?: string | null;
  /** @nullable */
  piTxid?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  shippingAddress?: string;
  notes?: string;
}

export type OrderStatusPatchStatus = typeof OrderStatusPatchStatus[keyof typeof OrderStatusPatchStatus];


export const OrderStatusPatchStatus = {
  pending: 'pending',
  paid: 'paid',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface OrderStatusPatch {
  status: OrderStatusPatchStatus;
  notes?: string;
}

export interface PaymentInitInput {
  orderId: number;
  amount: number;
  memo?: string;
}

export interface PaymentCompleteInput {
  paymentId: string;
  txid: string;
  orderId?: number;
}

export interface PaymentCancelInput {
  paymentId: string;
  orderId?: number;
}

export interface PaymentResult {
  success: boolean;
  status: string;
  /** @nullable */
  paymentId?: string | null;
  /** @nullable */
  txid?: string | null;
  /** @nullable */
  orderId?: number | null;
  /** @nullable */
  message?: string | null;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DailyStat {
  date: string;
  revenue: number;
  orders: number;
}

export interface AdminStats {
  totalUsers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  activeProducts: number;
  recentOrders?: Order[];
  ordersByStatus?: StatusCount[];
  revenueByDay?: DailyStat[];
}

export type ListProductsParams = {
category?: string;
search?: string;
page?: number;
limit?: number;
minPrice?: number;
maxPrice?: number;
sortBy?: string;
};

export type ListOrdersParams = {
status?: string;
};

export type ListAllOrdersParams = {
status?: string;
page?: number;
};