# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RAFMARKET - Pi Network Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/pi:
    post:
      operationId: authenticateWithPi
      tags: [auth]
      summary: Authenticate with Pi Network
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PiAuthInput"
      responses:
        "200":
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # CATEGORIES
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # PRODUCTS
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemPatch"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusPatch"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # PAYMENTS (Pi Network)
  /payments/pi/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a Pi Network payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
  /payments/pi/complete:
    post:
      operationId: completePayment
      tags: [payments]
      summary: Complete a Pi Network payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCompleteInput"
      responses:
        "200":
          description: Payment completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
  /payments/pi/cancel:
    post:
      operationId: cancelPayment
      tags: [payments]
      summary: Cancel a Pi Network payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCancelInput"
      responses:
        "200":
          description: Payment cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"

  # ADMIN
...
components:
  schemas:
...
    PiAuthInput:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, piUsername, role, createdAt]
      properties:
        id:
          type: integer
        piUsername:
          type: string
        piUid:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin, seller]
        createdAt:
          type: string
          format: date-time

    # Categories
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
    CategoryPatch:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string

    # Products
...
    CartItem:
      type: object
      required: [productId, productName, price, currency, quantity, imageUrl]
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        currency:
          type: string
        quantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
    Cart:
      type: object
      required: [items, total, currency]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        currency:
          type: string
        itemCount:
          type: integer
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    CartItemPatch:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # Orders
    OrderItem:
      type: object
      required: [productId, productName, price, currency, quantity]
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        currency:
          type: string
        quantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, userId, items, total, currency, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        buyerUsername:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled, refunded]
        piPaymentId:
          type: ["string", "null"]
        piTxid:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      properties:
        shippingAddress:
          type: string
        notes:
          type: string
    OrderStatusPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled, refunded]
        notes:
          type: string

    # Payments
    PaymentInitInput:
      type: object
      required: [orderId, amount]
      properties:
        orderId:
          type: integer
        amount:
          type: number
        memo:
          type: string
    PaymentCompleteInput:
      type: object
      required: [paymentId, txid]
      properties:
        paymentId:
          type: string
        txid:
          type: string
        orderId:
          type: integer
    PaymentCancelInput:
      type: object
      required: [paymentId]
      properties:
        paymentId:
          type: string
        orderId:
          type: integer
    PaymentResult:
      type: object
      required: [success, status]
      properties:
        success:
          type: boolean
        status:
          type: string
        paymentId:
          type: ["string", "null"]
        txid:
          type: ["string", "null"]
        orderId:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]

    # Admin
...