# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RAFMARKET - Pi Network Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  currency: string;
  quantity: number;
  /** @nullable */
  imageUrl: string | null;
}