# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RAFMARKET - Pi Network Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  buyerUsername?: string | null;
  items: OrderItem[];
  total: number;
  currency: string;
  status: OrderStatus;
  /** @nullable */
  piPaymentId?: string | null;
  /** @nullable */
  piTxid?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}