# lib/api-zod/src/generated/types/orderStatusPatch.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RAFMARKET - Pi Network Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusPatchStatus } from './orderStatusPatchStatus';

export interface OrderStatusPatch {
  status: OrderStatusPatchStatus;
  notes?: string;
}